i18n, unused component deletion, prettier timeout
