+++ public/locales/en/common.json
@@ -5,5 +5,6 @@
   "select": "Select",
   "swap": "Swap",
   "fetching": "Fetching",
-  "token": "Token"
+  "token": "Token",
+  "back": "Back"
 }
