+++ public/locales/en/marketing.json
@@ -1,0 +1,4 @@
+{
+  "zeroSlippage": "Zero slippage atomic swaps",
+  "communityFeeDiscout": "Low fees for community members"
+}
