+++ public/locales/en/orders.json
@@ -1,6 +1,7 @@
 {
   "continue": "Continue",
   "approve": "Approve",
+  "approvalRequired": "To continue you must approve {{symbol}}",
   "approve_aria": "Approve token spend",
   "take": "Complete Swap",
   "take_aria": "Take swap",
@@ -14,12 +15,15 @@
   "priceDifference": "Price difference",
   "bestQuote": "Best quote",
   "receivedQuotes": "Received quotes",
-  "bestPrice": "Best price",
+  "findingBestPrice": "Finding the best prices...",
+  "scanningPeers": "Scanning peers on the network",
   "price": "Price",
   "gasCost": "Gas cost",
   "protocolFee": "Protocol Fee",
   "searchByNameOrSymbol": "Search by name or symbol...",
   "removeFromActive": "Remove from active",
   "expandedResults": "Expanded results from inactive Token Lists",
-  "newQuoteIn": "New Quote In"
+  "newQuoteIn": "New Quote In",
+  "protocolFeeActive": "{{fee}} protocol fee active",
+  "afterFee": "After {{fee}} fee"
 }
