+++ src/components/TokenSelect/TokenSelect.stories.tsx
@@ -1,7 +1,7 @@
 import { Story, Meta } from "@storybook/react";
 import { TokenInfo } from "@uniswap/token-lists";
 
-import TokenSelect, { TokenSelectProps } from "./TokenSelect2";
+import TokenSelect, { TokenSelectProps } from "./TokenSelect";
 
 const wethTokenInfo: TokenInfo = {
   address: "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
