+++ src/components/TokenSelect/TokenSelect.tsx
@@ -1,98 +1,115 @@
-import { useEffect, useState } from "react";
+import { FC, MouseEventHandler, FormEventHandler } from "react";
 import { useTranslation } from "react-i18next";
 
-import { Web3Provider } from "@ethersproject/providers";
 import { TokenInfo } from "@uniswap/token-lists";
-import { useWeb3React } from "@web3-react/core";
 
+import TokenLogo from "../TokenLogo/TokenLogo";
 import {
-  StyledTextInput,
-  StyledTokenSelect,
-  TokenSelectorButton,
-  TokenSelectorLoader,
+  AmountInput,
+  AmountAndDetailsContainer,
+  AmountSubtext,
+  FlexRow,
+  TokenSelectContainer,
+  StyledSelectItem,
+  StyledSelectButton,
+  PlaceholderContainer,
+  PlaceholderTop,
+  PlaceholderBottom,
+  StyledLabel,
+  StyledDownArrow,
+  StyledSelectButtonContent,
 } from "./TokenSelect.styles";
 
 export type TokenSelectProps = {
-  withAmount: boolean;
-  quoteAmount?: string | null;
+  /**
+   * Whether or not the token selector is read only, e.g. when used to display
+   * a quote.
+   */
+  readOnly: boolean;
+  /**
+   * Whether or not to display an amount input with the token select, or to
+   * display a placeholder, e.g. when used to select received token where amount
+   * isn't known yet.
+   */
+  includeAmountInput: boolean;
+  /**
+   * Label - note that this is not shown if the field is readOnly.
+   */
   label: string;
-  readOnly?: boolean;
-  className?: string;
-  amount?: string;
-  token?: string;
-  tokens: TokenInfo[];
-  hasError?: boolean;
-  onAmountChange?: React.FormEventHandler<HTMLInputElement>;
-  onTokenChange?: () => void;
+  /**
+   * Metadata for currently selected token
+   */
+  selectedToken: TokenInfo | null;
+  /**
+   * Called when the user has clicked on the token dropdown to change token
+   */
+  onChangeTokenClicked: MouseEventHandler<HTMLButtonElement>;
+  /**
+   * Currently selected amount. Not used if `includeAmountInput` is false.
+   */
+  amount?: string | null;
+  /**
+   * Called when the amount is changed. Input change event is passed.
+   */
+  onAmountChange?: FormEventHandler<HTMLInputElement>;
+  /**
+   * Text to include beneath the amount. Currently used for fee information.
+   */
+  amountDetails?: string;
 };
 
-const TokenSelect = ({
-  withAmount,
-  quoteAmount,
+const TokenSelect: FC<TokenSelectProps> = ({
+  readOnly,
+  includeAmountInput,
   label,
-  className,
-  tokens,
+  selectedToken,
+  onChangeTokenClicked,
   amount,
   onAmountChange,
-  token,
-  onTokenChange,
-  hasError = false,
-}: TokenSelectProps) => {
-  const { t } = useTranslation(["common", "orders"]);
-  const { chainId } = useWeb3React<Web3Provider>();
-  const [isDefaultOptionDisabled, setIsDefaultOptionDisabled] = useState(false);
-  const [tokenSelectorText, setTokenSelectorText] = useState("");
-
-  useEffect(() => {
-    setIsDefaultOptionDisabled(false);
-  }, [chainId]);
-
-  useEffect(() => {
-    const match = tokens.find((t) => t.address === token)?.symbol;
-    const text =
-      match || isDefaultOptionDisabled ? match : `${t("common:select")}…`;
-
-    setTokenSelectorText(text || "");
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [tokens, token, isDefaultOptionDisabled]);
-
+  amountDetails,
+}) => {
+  const { t } = useTranslation(["common"]);
   return (
-    <StyledTokenSelect className={className} hasToken={!!token}>
-      {withAmount ? (
-        <StyledTextInput
-          label={label}
-          inputMode="decimal"
-          autoComplete="off"
-          pattern="^[0-9]*[.,]?[0-9]*$"
-          minLength={1}
-          maxLength={79}
-          spellCheck={false}
-          value={amount}
-          hasError={hasError}
-          onChange={onAmountChange}
-          placeholder="0.0"
-        />
+    <TokenSelectContainer>
+      <FlexRow>
+        <TokenLogo size="medium" tokenInfo={selectedToken} />
+        <StyledSelectButton onClick={onChangeTokenClicked} disabled={readOnly}>
+          <StyledLabel invisible={readOnly}>{label}</StyledLabel>
+          <StyledSelectItem>
+            <StyledSelectButtonContent emphasize={readOnly}>
+              {selectedToken !== null
+                ? selectedToken.symbol
+                : t("common:select")}
+            </StyledSelectButtonContent>
+            <StyledDownArrow invisible={readOnly} />
+          </StyledSelectItem>
+        </StyledSelectButton>
+      </FlexRow>
+      {includeAmountInput ? (
+        <AmountAndDetailsContainer>
+          <AmountInput
+            // @ts-ignore
+            inputMode="decimal"
+            autoComplete="off"
+            pattern="^[0-9]*[.,]?[0-9]*$"
+            minLength={1}
+            maxLength={79}
+            spellCheck={false}
+            value={amount}
+            disabled={readOnly}
+            onChange={onAmountChange}
+            placeholder="0.00"
+            hasSubtext={!!amountDetails}
+          />
+          {amountDetails && <AmountSubtext>{amountDetails}</AmountSubtext>}
+        </AmountAndDetailsContainer>
       ) : (
-        <StyledTextInput
-          disabled
-          label={label}
-          autoComplete="off"
-          spellCheck={false}
-          value={!token ? t("orders:chooseToken") : quoteAmount || ""}
-        />
+        <PlaceholderContainer>
+          <PlaceholderTop />
+          <PlaceholderBottom />
+        </PlaceholderContainer>
       )}
-
-      {tokens ? (
-        <TokenSelectorButton
-          iconSize={0.675}
-          icon="arrow-right"
-          text={tokenSelectorText}
-          onClick={() => onTokenChange && onTokenChange()}
-        />
-      ) : (
-        <TokenSelectorLoader />
-      )}
-    </StyledTokenSelect>
+    </TokenSelectContainer>
   );
 };
