+++ src/i18n/react-i18n.d.ts
@@ -4,6 +4,7 @@ import "react-i18next";
 // import all namespaces (for the default language, only)
 import balances from "../../public/locales/en/balances.json";
 import common from "../../public/locales/en/common.json";
+import marketing from "../../public/locales/en/marketing.json";
 import orders from "../../public/locales/en/orders.json";
 import wallet from "../../public/locales/en/wallet.json";
 
@@ -14,5 +15,6 @@ declare module "react-i18next" {
     common: typeof common;
     wallet: typeof wallet;
     orders: typeof orders;
+    marketing: typeof marketing;
   }
 }
