+++ src/prettier.spec.ts
@@ -1,5 +1,8 @@
 import { exec } from "child_process";
 
+// This test can take a long time to run, and that's okay!
+jest.setTimeout(30000);
+
 describe("Prettier conformance", () => {
   it("all files should have been formatted by prettier", (done) => {
     exec(
